S1C8 - Height Consideration & Pointing & Clicking Functionality
Game now places the game environment at an appropraite height based on the height of the camera on the start. This will eventually change based on when the player puts the headset on. Pointer can now point and click at objects. This highlights them, and also returns the object itself for use with interactions later on.