S4C12 - Rotation Removed
Physics based rotation was too intensive for the possibility of lots of game pieces on the board, so an alternative was tried not using physics but it didn't feel intuitive enough to warrant having it as a feature. If the player wishes to rotate the board they should just walk around it.