S14C3 - Timer Sound Fixed + Building Destruction Effect + Ability to skip building
Timer sound has been fixes so it happens on the tick. Building destruction particle effect added. Ability to skip the building phase has also been added, on the same button that turns it to "Double speed" whilst the player is in the the attack phase. This swaps out as the player goes through the phases